some tests created a TokenIo but just needed Util; so just use Util
When we added devKey, we overlooked adding it in some tests--but
those tests, confusingly, still passed.

This was because those tests created a TokenIo, but only ever
used TokenIo.Util.generateNonce from it.

To avoid confusing ourselves more in the future, modified
those tests to just include Util instead of making
SDK clients.