ease understanding of CryptoEngine interface
pull some "common code" out of crypto engines and put it
into Crypto.js. It wasn't super-boiler-plate-y; but this
way seems easier to understand "we expect to you customize
THIS part; we expect you to do THIS part in 'the usual way'"

Add a README.md file that spells out what the
"CryptoEngine interface" means in this duck-type-y JS world.