member method samples (#65)
sample code for some simple Member methods. In langs with a type system, we don't have sample code for these because they're basically "fill in a struct, call member.set( struct )" . but for JS, "fill in a struct" is not so obvious.

bonus comment fix: comments said address in some places where it's now AddressRecord.